+++ malachite-float/src/bin_util/demo_and_bench/float/arithmetic/cos.rs
@@ -15,7 +15,7 @@ use malachite_base::test_util::bench::{BenchmarkType, run_benchmark};
 use malachite_base::test_util::generators::common::{GenConfig, GenMode};
 use malachite_base::test_util::generators::primitive_float_gen;
 use malachite_base::test_util::runner::Runner;
-use malachite_float::float::arithmetic::cos::primitive_float_cos;
+use malachite_float::float::arithmetic::cos::{primitive_float_cos, primitive_float_cos_rational};
 use malachite_float::test_util::bench::bucketers::{
     float_complexity_bucketer, pair_2_float_complexity_bucketer,
     pair_2_triple_1_2_float_primitive_int_max_complexity_bucketer,
@@ -27,8 +27,14 @@ use malachite_float::test_util::generators::{
     float_unsigned_pair_gen_var_1, float_unsigned_pair_gen_var_4,
     float_unsigned_rounding_mode_triple_gen_var_36,
     float_unsigned_rounding_mode_triple_gen_var_36_rm,
+    rational_unsigned_rounding_mode_triple_gen_var_10,
 };
 use malachite_float::{ComparableFloat, ComparableFloatRef, Float};
+use malachite_q::test_util::bench::bucketers::{
+    pair_rational_bit_u64_max_bucketer, rational_bit_bucketer,
+    triple_1_2_rational_bit_u64_max_bucketer,
+};
+use malachite_q::test_util::generators::{rational_gen, rational_unsigned_pair_gen_var_3};
 
 pub(crate) fn register(runner: &mut Runner) {
     register_demo!(runner, demo_float_cos);
@@ -49,6 +55,15 @@ pub(crate) fn register(runner: &mut Runner) {
     register_demo!(runner, demo_float_cos_round_ref);
     register_demo!(runner, demo_float_cos_round_assign);
     register_primitive_float_demos!(runner, demo_primitive_float_cos);
+    register_primitive_float_demos!(runner, demo_primitive_float_cos_rational);
+    register_demo!(runner, demo_float_cos_rational_prec);
+    register_demo!(runner, demo_float_cos_rational_prec_debug);
+    register_demo!(runner, demo_float_cos_rational_prec_ref);
+    register_demo!(runner, demo_float_cos_rational_prec_ref_debug);
+    register_demo!(runner, demo_float_cos_rational_prec_round);
+    register_demo!(runner, demo_float_cos_rational_prec_round_debug);
+    register_demo!(runner, demo_float_cos_rational_prec_round_ref);
+    register_demo!(runner, demo_float_cos_rational_prec_round_ref_debug);
     register_demo!(runner, demo_float_cos_prec_round);
     register_demo!(runner, demo_float_cos_prec_round_debug);
     register_demo!(runner, demo_float_cos_prec_round_ref);
@@ -60,6 +75,15 @@ pub(crate) fn register(runner: &mut Runner) {
     register_bench!(runner, benchmark_float_cos_prec_round_evaluation_strategy);
     register_bench!(runner, benchmark_float_cos_prec_round_library_comparison);
     register_primitive_float_benches!(runner, benchmark_primitive_float_cos);
+    register_primitive_float_benches!(runner, benchmark_primitive_float_cos_rational);
+    register_bench!(
+        runner,
+        benchmark_float_cos_rational_prec_evaluation_strategy
+    );
+    register_bench!(
+        runner,
+        benchmark_float_cos_rational_prec_round_evaluation_strategy
+    );
 }
 
 fn demo_float_cos(gm: GenMode, config: &GenConfig, limit: usize) {
@@ -434,6 +458,226 @@ fn benchmark_primitive_float_cos<T: PrimitiveFloat>(
         &primitive_float_bucketer("x"),
         &mut [("malachite", &mut |x| {
             no_out!(primitive_float_cos(x));
+        })],
+    );
+}
+
+fn demo_float_cos_rational_prec(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (n, p) in rational_unsigned_pair_gen_var_3()
+        .get(gm, config)
+        .take(limit)
+    {
+        println!(
+            "Float::cos_rational_prec({}, {}) = {:?}",
+            n.clone(),
+            p,
+            Float::cos_rational_prec(n, p)
+        );
+    }
+}
+
+fn demo_float_cos_rational_prec_debug(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (n, p) in rational_unsigned_pair_gen_var_3()
+        .get(gm, config)
+        .take(limit)
+    {
+        let (f, o) = Float::cos_rational_prec(n.clone(), p);
+        println!(
+            "Float::cos_rational_prec({}, {}) = ({:#x}, {:?})",
+            n,
+            p,
+            ComparableFloat(f),
+            o
+        );
+    }
+}
+
+fn demo_float_cos_rational_prec_ref(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (n, p) in rational_unsigned_pair_gen_var_3()
+        .get(gm, config)
+        .take(limit)
+    {
+        println!(
+            "Float::cos_rational_prec_ref(&{}, {}) = {:?}",
+            n,
+            p,
+            Float::cos_rational_prec_ref(&n, p)
+        );
+    }
+}
+
+fn demo_float_cos_rational_prec_ref_debug(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (n, p) in rational_unsigned_pair_gen_var_3()
+        .get(gm, config)
+        .take(limit)
+    {
+        let (f, o) = Float::cos_rational_prec_ref(&n, p);
+        println!(
+            "Float::cos_rational_prec_ref(&{}, {}) = ({:#x}, {:?})",
+            n,
+            p,
+            ComparableFloat(f),
+            o
+        );
+    }
+}
+
+fn demo_float_cos_rational_prec_round(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (n, p, rm) in rational_unsigned_rounding_mode_triple_gen_var_10()
+        .get(gm, config)
+        .take(limit)
+    {
+        println!(
+            "Float::cos_rational_prec_round({}, {}, {}) = {:?}",
+            n.clone(),
+            p,
+            rm,
+            Float::cos_rational_prec_round(n, p, rm)
+        );
+    }
+}
+
+fn demo_float_cos_rational_prec_round_debug(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (n, p, rm) in rational_unsigned_rounding_mode_triple_gen_var_10()
+        .get(gm, config)
+        .take(limit)
+    {
+        let (f, o) = Float::cos_rational_prec_round(n.clone(), p, rm);
+        println!(
+            "Float::cos_rational_prec_round({}, {}, {}) = ({:#x}, {:?})",
+            n,
+            p,
+            rm,
+            ComparableFloat(f),
+            o
+        );
+    }
+}
+
+fn demo_float_cos_rational_prec_round_ref(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (n, p, rm) in rational_unsigned_rounding_mode_triple_gen_var_10()
+        .get(gm, config)
+        .take(limit)
+    {
+        println!(
+            "Float::cos_rational_prec_round_ref(&{}, {}, {}) = {:?}",
+            n,
+            p,
+            rm,
+            Float::cos_rational_prec_round_ref(&n, p, rm)
+        );
+    }
+}
+
+fn demo_float_cos_rational_prec_round_ref_debug(gm: GenMode, config: &GenConfig, limit: usize) {
+    for (n, p, rm) in rational_unsigned_rounding_mode_triple_gen_var_10()
+        .get(gm, config)
+        .take(limit)
+    {
+        let (f, o) = Float::cos_rational_prec_round_ref(&n, p, rm);
+        println!(
+            "Float::cos_rational_prec_round_ref(&{}, {}, {}) = ({:#x}, {:?})",
+            n,
+            p,
+            rm,
+            ComparableFloat(f),
+            o
+        );
+    }
+}
+
+fn benchmark_float_cos_rational_prec_evaluation_strategy(
+    gm: GenMode,
+    config: &GenConfig,
+    limit: usize,
+    file_name: &str,
+) {
+    run_benchmark(
+        "Float::cos_rational_prec(Rational, u64)",
+        BenchmarkType::EvaluationStrategy,
+        rational_unsigned_pair_gen_var_3().get(gm, config),
+        gm.name(),
+        limit,
+        file_name,
+        &pair_rational_bit_u64_max_bucketer("n", "prec"),
+        &mut [
+            (
+                "Float::cos_rational_prec(Rational, u64)",
+                &mut |(n, prec)| no_out!(Float::cos_rational_prec(n, prec)),
+            ),
+            (
+                "Float::cos_rational_prec_ref(&Rational, u64)",
+                &mut |(n, prec)| no_out!(Float::cos_rational_prec_ref(&n, prec)),
+            ),
+        ],
+    );
+}
+
+fn benchmark_float_cos_rational_prec_round_evaluation_strategy(
+    gm: GenMode,
+    config: &GenConfig,
+    limit: usize,
+    file_name: &str,
+) {
+    run_benchmark(
+        "Float::cos_rational_prec_round(Rational, u64, RoundingMode)",
+        BenchmarkType::EvaluationStrategy,
+        rational_unsigned_rounding_mode_triple_gen_var_10().get(gm, config),
+        gm.name(),
+        limit,
+        file_name,
+        &triple_1_2_rational_bit_u64_max_bucketer("n", "prec"),
+        &mut [
+            (
+                "Float::cos_rational_prec_round(Rational, u64, RoundingMode)",
+                &mut |(n, prec, rm)| no_out!(Float::cos_rational_prec_round(n, prec, rm)),
+            ),
+            (
+                "Float::cos_rational_prec_round_ref(&Rational, u64, RoundingMode)",
+                &mut |(n, prec, rm)| no_out!(Float::cos_rational_prec_round_ref(&n, prec, rm)),
+            ),
+        ],
+    );
+}
+
+#[allow(clippy::type_repetition_in_bounds)]
+fn demo_primitive_float_cos_rational<T: PrimitiveFloat>(
+    gm: GenMode,
+    config: &GenConfig,
+    limit: usize,
+) where
+    Float: From<T> + PartialOrd<T>,
+    for<'a> T: ExactFrom<&'a Float> + RoundingFrom<&'a Float>,
+{
+    for x in rational_gen().get(gm, config).take(limit) {
+        println!(
+            "primitive_float_cos_rational({}) = {:?}",
+            x,
+            NiceFloat(primitive_float_cos_rational::<T>(&x))
+        );
+    }
+}
+
+#[allow(clippy::type_repetition_in_bounds)]
+fn benchmark_primitive_float_cos_rational<T: PrimitiveFloat>(
+    gm: GenMode,
+    config: &GenConfig,
+    limit: usize,
+    file_name: &str,
+) where
+    Float: From<T> + PartialOrd<T>,
+    for<'a> T: ExactFrom<&'a Float> + RoundingFrom<&'a Float>,
+{
+    run_benchmark(
+        &format!("primitive_float_cos_rational::<{}>(&Rational)", T::NAME),
+        BenchmarkType::Single,
+        rational_gen().get(gm, config),
+        gm.name(),
+        limit,
+        file_name,
+        &rational_bit_bucketer("x"),
+        &mut [("malachite", &mut |x| {
+            no_out!(primitive_float_cos_rational::<T>(&x));
         })],
     );
 }