+++ malachite-float/src/test_util/generators/mod.rs
@@ -5665,7 +5665,8 @@ pub fn rational_unsigned_rounding_mode_triple_gen_var_9() -> GT3 {
     )
 }
 
-// All `(Rational, u64, RoundingMode)` that are valid inputs to `Float::exp_rational_prec_round`.
+// All `(Rational, u64, RoundingMode)` that are valid inputs to `Float::exp_rational_prec_round` and
+// `Float::cos_rational_prec_round`.
 pub fn rational_unsigned_rounding_mode_triple_gen_var_10() -> GT3 {
     Generator::new(
         &exhaustive_rational_unsigned_rounding_mode_triple_gen_var_10,